requirements.txt
numpy==1.19.5
Pillow==10.3.0
scikit_learn==1.4.2
streamlit==1.33.0
tensorflow==2.4.1
tensorflow_intel==2.13.0
tqdm==4.66.2
